Implement on section user auth

+++ src/main/java/com/taskflow/studentmanagement/controller/InstructorController.java
@@ -10,8 +10,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.taskflow.studentmanagement.exception.ResourceNotFoundException;
-import com.taskflow.studentmanagement.io.request.InstructorRequest;
-import com.taskflow.studentmanagement.io.response.InstructorResponse;
+import com.taskflow.studentmanagement.dto.request.InstructorRequest;
+import com.taskflow.studentmanagement.dto.response.InstructorResponse;
 import com.taskflow.studentmanagement.service.InstructorService;
 
 import jakarta.validation.Valid;
@@ -30,10 +30,11 @@ public class InstructorController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<InstructorResponse> getById(@PathVariable Long id) {
+    public ResponseEntity<InstructorResponse> getById(@PathVariable Integer id) {
         return ResponseEntity.ok(instructorService.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Instructor or Prof. not found"))
         );
     }
+
 }
